requirements.txt
accelerate==0.31.0
diffusers==0.27.2
gradio==3.39.0
gradio-client==1.3.0
h5py==3.11.0
huggingface-hub==0.24.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.4.0
importlib-resources==6.4.5
ipdb==0.13.13
ipython==8.26.0
ipywidgets==8.1.5
kornia==0.7.3
kornia-rs==0.1.5
omegaconf==2.3.0
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.1
torchvision==0.18.0
tqdm==4.66.5
transformers==4.44.1
einops==0.8.0
basicsr==1.3.5